requirement.txt
Django==1.10.3
image==1.5.5
Pillow==3.4.2
pyobjc-core==3.1.1
pyobjc-framework-Cocoa==3.1.1
pyobjc-framework-Quartz==3.1.1
wheel==0.24.0
